Separate result pages and search suggestions
To make navigation using browser history more
user friendly, e.g. when navigating back to the
page after clicking a link to TMDB, the user will
see the search results on the correct page they
left from. To do this, the server just returns a
new page instead of just updating the current page
with an ajax call.

Also added page numbers, next, and previous page links
to the bottom of the page.

Finally, as the user types a search, we now query TMDB
to provide a completion list of the closest matching
titles.